Main is default dispatcher in rememberLocalStore
We want to utilize Kotlin's suspend functions in the side effects. Therefor it makes sense to use the Main dispatcher for running side effects, because suspend functions should be safe to call from the main thread. In this case it makes the Default dispatcher redundant.
This will move the responsibility of changing thread/dispatcher to suspend functions called in the side effects or libraries that support suspend functions like Retrofit and Room.

ApiClient in the sample app was updated to use suspend functions to simulator real usages of suspend functions.